lib/libc: Replace long => uintptr_t in mem{cpy,move}
The lk implementations of mem{cpy,move} internally
cast some pointers to long, which triggers UBSan.
This replaces the target type with uintptr_t which
fixes the issue (and it matches the original char*
pointer types better).

Bug: 231151995

+++ lib/libc/string/memcpy.c
@@ -44,12 +44,12 @@ void *memcpy(void *dest, const void *src, size_t count)
     if (count == 0 || dest == src)
         return dest;
 
-    if (((long)d | (long)s) & lmask) {
+    if (((uintptr_t)d | (uintptr_t)s) & lmask) {
         // src and/or dest do not align on word boundary
-        if ((((long)d ^ (long)s) & lmask) || (count < lsize))
+        if ((((uintptr_t)d ^ (uintptr_t)s) & lmask) || (count < lsize))
             len = count; // copy the rest of the buffer with the byte mover
         else
-            len = lsize - ((long)d & lmask); // move the ptrs up to a word boundary
+            len = lsize - ((uintptr_t)d & lmask); // move the ptrs up to a word boundary
 
         count -= len;
         for (; len > 0; len--)
